+++ s1tbx-io/src/main/java/org/esa/nest/dataio/radarsat2/Radarsat2ProductDirectory.java
@@ -504,8 +504,8 @@ public class Radarsat2ProductDirectory extends XMLProductDirectory {
 
         final MetadataElement[] geoGrid = geolocationGrid.getElements();
 
-        double[] latList = new double[geoGrid.length];
-        double[] lngList = new double[geoGrid.length];
+        float[] latList = new float[geoGrid.length];
+        float[] lngList = new float[geoGrid.length];
 
         int gridWidth = 0, gridHeight = 0;
         int i = 0;
@@ -513,8 +513,8 @@ public class Radarsat2ProductDirectory extends XMLProductDirectory {
             final MetadataElement geodeticCoordinate = imageTiePoint.getElement("geodeticCoordinate");
             final MetadataElement latitude = geodeticCoordinate.getElement("latitude");
             final MetadataElement longitude = geodeticCoordinate.getElement("longitude");
-            latList[i] = latitude.getAttributeDouble("latitude", 0);
-            lngList[i] = longitude.getAttributeDouble("longitude", 0);
+            latList[i] = (float)latitude.getAttributeDouble("latitude", 0);
+            lngList[i] = (float)longitude.getAttributeDouble("longitude", 0);
 
             final MetadataElement imageCoordinate = imageTiePoint.getElement("imageCoordinate");
             final double pix = imageCoordinate.getAttributeDouble("pixel", 0);
@@ -528,8 +528,8 @@ public class Radarsat2ProductDirectory extends XMLProductDirectory {
         }
 
         if (flipToSARGeometry) {
-            double[] flippedLatList = new double[geoGrid.length];
-            double[] flippedLonList = new double[geoGrid.length];
+            float[] flippedLatList = new float[geoGrid.length];
+            float[] flippedLonList = new float[geoGrid.length];
             int is, id;
             if (isAscending) {
                 if (isAntennaPointingRight) { // flip upside down
@@ -666,7 +666,7 @@ public class Radarsat2ProductDirectory extends XMLProductDirectory {
         final double theta1 = Math.acos((r1 + rt * Math.cos(alpha1)) / rtPlusH);
         final double psi1 = alpha1 - theta1;
         double psi = psi1;
-        double[] incidenceAngles = new double[gridWidth];
+        float[] incidenceAngles = new float[gridWidth];
         final int n = gridWidth * subSamplingX;
         int k = 0;
         for (int i = 0; i < n; i++) {
@@ -689,7 +689,7 @@ public class Radarsat2ProductDirectory extends XMLProductDirectory {
             psi = psi + deltaPsi;
         }
 
-        double[] incidenceAngleList = new double[gridWidth * gridHeight];
+        float[] incidenceAngleList = new float[gridWidth * gridHeight];
         for (int j = 0; j < gridHeight; j++) {
             System.arraycopy(incidenceAngles, 0, incidenceAngleList, j * gridWidth, gridWidth);
         }
@@ -746,8 +746,8 @@ public class Radarsat2ProductDirectory extends XMLProductDirectory {
         final int sceneHeight = product.getSceneRasterHeight();
         final int subSamplingX = sceneWidth / (gridWidth - 1);
         final int subSamplingY = sceneHeight / (gridHeight - 1);
-        final double[] rangeDist = new double[gridWidth * gridHeight];
-        final double[] rangeTime = new double[gridWidth * gridHeight];
+        final float[] rangeDist = new float[gridWidth * gridHeight];
+        final float[] rangeTime = new float[gridWidth * gridHeight];
 
         final coefList[] segments = segmentsArray.toArray(new coefList[segmentsArray.size()]);
 
@@ -775,7 +775,7 @@ public class Radarsat2ProductDirectory extends XMLProductDirectory {
                 final double g2 = g * g;
 
                 //SlantRange = s0 + s1(GR - GR0) + s2(GR-GR0)^2 + s3(GRGR0)^3 + s4(GR-GR0)^4;
-                rangeDist[k++] = s0 + s1 * g + s2 * g2 + s3 * g2 * g + s4 * g2 * g2;
+                rangeDist[k++] = (float)(s0 + s1 * g + s2 * g2 + s3 * g2 * g + s4 * g2 * g2);
             }
         }
 
@@ -785,7 +785,7 @@ public class Radarsat2ProductDirectory extends XMLProductDirectory {
             if (!flipToSARGeometry && (isDescending && isAntennaPointingRight || !isDescending && !isAntennaPointingRight)) // flip for descending RS2
                 index = rangeDist.length - 1 - i;
 
-            rangeTime[index] = rangeDist[i] / Constants.halfLightSpeed * Constants.oneBillion; // in ns
+            rangeTime[index] = (float)(rangeDist[i] / Constants.halfLightSpeed * Constants.oneBillion); // in ns
         }
 
         final TiePointGrid slantRangeGrid = new TiePointGrid(OperatorUtils.TPG_SLANT_RANGE_TIME,
